Change background color of float menu
The previous background color of the floating menu was the same as the
main background color. This made it difficult to distinguish between the
floating menu and the buffer behind it.

The change to the 'selection_inactive' color not only fixes that
visibility issue, but also brings it in line with the completion menu
highlight ('PMenu'), which makes sense, given that the floating menu
often appears in conjunction with the completion menu